Provide a way to customize NVM installation settings.

+++ src/main/java/org/jenkinsci/plugins/nvm/NvmWrapperUtil.java
@@ -5,6 +5,7 @@ import hudson.FilePath;
 import hudson.Launcher;
 import hudson.model.TaskListener;
 import hudson.util.ArgumentListBuilder;
+import org.apache.commons.lang.StringUtils;
 
 import java.io.File;
 import java.io.IOException;
@@ -18,49 +19,54 @@ public class NvmWrapperUtil {
   private Launcher launcher;
   private TaskListener listener;
 
-  private static final List<String> NVM_PATHS = Arrays.asList(
-    "~/.nvm/nvm.sh",
-    "/usr/local/nvm/nvm.sh"
-  );
-
   NvmWrapperUtil(final FilePath workspace, Launcher launcher, TaskListener listener) {
     this.workspace = workspace;
     this.listener = listener;
     this.launcher = launcher;
   }
 
-  public Map<String, String> getNpmEnvVars(String version, String nvmInstallURL,
-                                           String nvmNodeJsOrgMirror, String nvmIoJsOrgMirror)
+  public Map<String, String> getNpmEnvVars(String nodeVersion, String nvmInstallURL,
+                                           String nvmNodeJsOrgMirror, String nvmIoJsOrgMirror,
+                                           String nvmInstallDir)
     throws IOException, InterruptedException {
 
-    int statusCode = installNvm(Optional.ofNullable(nvmInstallURL).orElse(NvmDefaults.nvmInstallURL));
-    if (statusCode != 0) {
-      throw new AbortException("Failed to install NVM");
+
+    String nvmDir = StringUtils.defaultIfEmpty(nvmInstallDir, NvmDefaults.nvmInstallDir);
+    nvmDir = nvmDir.endsWith("/") ? StringUtils.stripEnd(nvmDir, "/")  : nvmDir;
+    final String nvmFilePath = nvmDir + "/nvm.sh";
+
+    if(fileExist(nvmFilePath) == false){ //NVM is not installed
+      int statusCode = installNvm(StringUtils.defaultIfEmpty(nvmInstallURL, NvmDefaults.nvmInstallURL),
+        nvmDir, nodeVersion);
+
+      if (statusCode != 0) {
+        throw new AbortException("Failed to install NVM");
+      }
+
+    }else{
+      listener.getLogger().println("NVM is already installed\n");
     }
 
-    String mirrorBin = version.contains("iojs") ?
-      "NVM_IOJS_ORG_MIRROR=" + Optional.ofNullable(nvmIoJsOrgMirror).orElse(NvmDefaults.nvmIoJsOrgMirror) :
-      "NVM_NODEJS_ORG_MIRROR=" + Optional.ofNullable(nvmNodeJsOrgMirror).orElse(NvmDefaults.nvmNodeJsOrgMirror);
+    String mirrorBin = nodeVersion.contains("iojs") ?
+      "NVM_IOJS_ORG_MIRROR=" + StringUtils.defaultIfEmpty(nvmIoJsOrgMirror, NvmDefaults.nvmIoJsOrgMirror) :
+      "NVM_NODEJS_ORG_MIRROR=" + StringUtils.defaultIfEmpty(nvmNodeJsOrgMirror, NvmDefaults.nvmNodeJsOrgMirror);
 
-    ArgumentListBuilder beforeCmd = new ArgumentListBuilder();
-    beforeCmd.add("bash");
-    beforeCmd.add("-c");
-    beforeCmd.add("export > before.env");
 
-    Map<String, String> beforeEnv = toMap(getExport(beforeCmd, "before.env"));
+    Map<String, String> beforeEnv = getExport();
+
+    String envFile="env.txt";
 
     ArgumentListBuilder nvmSourceCmd = new ArgumentListBuilder();
     nvmSourceCmd.add("bash");
     nvmSourceCmd.add("-c");
     nvmSourceCmd.add(
-        NVM_PATHS.stream().map(
-          path -> "{ [ -f " + path + " ] && source " + path + "; }")
-          .collect(Collectors.joining(" || ")) +
-        " && " + mirrorBin + " nvm install " + version +
-        " && nvm use " + version +
-        " && export > nvm.env");
+        "NVM_DIR=" + nvmDir +
+        " && source $NVM_DIR/nvm.sh "+
+        " && " + mirrorBin +  " nvm install " + nodeVersion +
+        " && nvm use " + nodeVersion + " && export > " + envFile );
 
-    Map<String, String> afterEnv = toMap(getExport(nvmSourceCmd, "nvm.env"));
+
+    Map<String, String> afterEnv = getExport(nvmSourceCmd, envFile);
 
     Map<String, String> newEnvVars = new HashMap<>();
 
@@ -84,7 +90,9 @@ public class NvmWrapperUtil {
     return newEnvVars;
   }
 
-  private String getExport(ArgumentListBuilder args, String destFile) throws IOException, InterruptedException {
+
+
+  private Map<String, String> getExport(ArgumentListBuilder args, String destFile ) throws IOException, InterruptedException {
 
     Integer statusCode = launcher.launch().pwd(workspace).cmds(args)
       .stdout(listener.getLogger())
@@ -94,20 +102,69 @@ public class NvmWrapperUtil {
       throw new AbortException("Failed to fork bash ");
     }
 
-    return workspace.child(destFile).readToString();
+    String out = workspace.child(destFile).readToString();
+
+    workspace.child(destFile).delete();
+
+    return toMap(out);
+  }
+  private Map<String, String> getExport() throws IOException, InterruptedException {
+
+    String destFile="env.txt";
+    ArgumentListBuilder args = new ArgumentListBuilder();
+    args.add("bash");
+    args.add("-c");
+    args.add("export > " + destFile);
+    return getExport(args, destFile);
 
   }
 
-  private Integer installNvm(String nvmInstallURL) throws IOException, InterruptedException {
+  private Boolean fileExist(String filePath) throws IOException, InterruptedException {
+    ArgumentListBuilder args = new ArgumentListBuilder();
+      args.add("bash");
+      args.add("-c");
+      args.add("test -f " + filePath);
+
+    Integer statusCode = launcher.launch().pwd(workspace).cmds(args)
+      .stdout(listener.getLogger())
+      .stderr(listener.getLogger()).join();
+
+    return statusCode == 0;
+  }
+
+  private Integer installNvm(String nvmInstallURL, String nvmInstallDir,
+                             String nvmInstallNodeVersion) throws IOException, InterruptedException {
     listener.getLogger().println("Installing nvm\n");
     FilePath installer = workspace.child("nvm-installer");
     installer.copyFrom(new URL(nvmInstallURL));
     installer.chmod(0755);
+
     ArgumentListBuilder args = new ArgumentListBuilder();
+    args.add("bash");
+    args.add("-c");
 
-    args.add(installer.absolutize().getRemote());
 
-    return launcher.launch().cmds(args).pwd(workspace)
+    List<String> cmdBuild = new ArrayList<>();
+
+//    only add if default is changed
+//    https://github.com/creationix/nvm/blob/master/install.sh#L300
+//    nvm script just create install dir if it is the default.
+    if(!nvmInstallDir.equals(NvmDefaults.nvmInstallDir)){
+      cmdBuild.add("NVM_DIR=" + nvmInstallDir);
+    }
+
+
+
+    if (StringUtils.isNotBlank(nvmInstallNodeVersion)) {
+      cmdBuild.add("NODE_VERSION=" + nvmInstallNodeVersion);
+    }
+
+    cmdBuild.add("NVM_PROFILE=''"); //Avoid modifying profile
+    cmdBuild.add(installer.absolutize().getRemote());
+
+    args.add(cmdBuild.stream().collect(Collectors.joining(" ")));
+
+    return launcher.launch().pwd(workspace).cmds(args)
       .stdout(listener.getLogger())
       .stderr(listener.getLogger()).join();
